Fixing a bug in CountAttributeAggregator#restore method.

+++ siddhi-core/src/main/java/org/wso2/siddhi/core/query/selector/attribute/aggregator/CountAttributeAggregator.java
@@ -105,6 +105,6 @@ public class CountAttributeAggregator extends AttributeAggregator {
 
     @Override
     public void restoreState(Map<String, Object> state) {
-        count = (int) state.get("Count");
+        count = (long) state.get("Count");
     }
 }
